# docs/specs/openapi/dkim2d.yaml
openapi: 3.0.3
info:
  title: DKIM2 Daemon API
  version: 0.1.0-draft
  description: >
    Verification, signing, revision, policy, and replay API for the DKIM2
    reference daemon. Protocol behavior is pinned to
    draft-ietf-dkim-dkim2-spec-04.
paths:
...
  /v1/process:
    post:
      operationId: processMessage
      summary: Process one inbound DKIM2 message.
      description: >
        Verifies the current message instance, applies server-owned policy,
        performs the configured replay check, and returns a closed final
        disposition. This operation neither signs nor revises a message.
      security:
        - localCapability: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessRequest"
      responses:
        "200":
          description: Processing completed with a coherent domain result.
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/ContentTypeOptions"
            Content-Length:
              $ref: "#/components/headers/ContentLength"
            Connection:
              $ref: "#/components/headers/Connection"
            Date:
              $ref: "#/components/headers/Date"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "408":
          $ref: "#/components/responses/RequestTimeout"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "417":
          $ref: "#/components/responses/ExpectationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/sign:
    post:
      operationId: signMessage
      summary: Sign one originator DKIM2 message.
      description: >
        Resolves server-owned originator policy and signing material and
        returns a closed ordered append-only action plan.
      security:
        - localCapability: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignRequest"
      responses:
        "200":
          description: Originator signing completed with a coherent result.
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/ContentTypeOptions"
            Content-Length:
              $ref: "#/components/headers/ContentLength"
            Connection:
              $ref: "#/components/headers/Connection"
            Date:
              $ref: "#/components/headers/Date"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "408":
          $ref: "#/components/responses/RequestTimeout"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "417":
          $ref: "#/components/responses/ExpectationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/revise:
    post:
      operationId: reviseMessage
      summary: Revise one ordinary-transit DKIM2 message.
      description: >
        Verifies exact inherited revision evidence before server-owned
        ordinary-transit signing and returns a closed ordered action plan.
      security:
        - localCapability: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviseRequest"
      responses:
        "200":
          description: Ordinary-transit revision completed coherently.
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/ContentTypeOptions"
            Content-Length:
              $ref: "#/components/headers/ContentLength"
            Connection:
              $ref: "#/components/headers/Connection"
            Date:
              $ref: "#/components/headers/Date"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "408":
          $ref: "#/components/responses/RequestTimeout"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "417":
          $ref: "#/components/responses/ExpectationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  securitySchemes:
    localCapability:
      type: apiKey
      in: header
      name: X-DKIM2-Capability
      description: >
        Generation-bound local process capability. This credential is not an
        OAuth bearer token and is not intended for remote transport.
  headers:
    CacheControl:
      description: Prevent storage of every daemon response.
      required: true
      schema:
        type: string
        enum:
          - no-store
    ContentTypeOptions:
      description: Prevent content-type sniffing.
      required: true
      schema:
        type: string
        enum:
          - nosniff
    ContentLength:
      description: Canonical non-negative decimal representation length.
      required: true
      schema:
        type: string
        pattern: "^(0|[1-9][0-9]*)$"
...
    Connection:
      description: Close the accepted connection after this response.
      required: true
      schema:
        type: string
        enum:
          - close
    Date:
      description: RFC 9110 origin-server date when a valid clock is available.
      schema:
        type: string
        pattern: "^(Mon|Tue|Wed|Thu|Fri|Sat|Sun), [0-9]{2} (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec) [0-9]{4} [0-9]{2}:[0-9]{2}:[0-9]{2} GMT$"
    RetryAfter:
      description: Fixed retry delay in seconds.
      required: true
      schema:
        type: string
        enum:
          - "1"
  responses:
...
    BadRequest:
      description: The request is invalid.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
        Content-Length:
          $ref: "#/components/headers/ContentLength"
        Connection:
          $ref: "#/components/headers/Connection"
        Date:
          $ref: "#/components/headers/Date"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The local process capability is absent or invalid.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
        Content-Length:
          $ref: "#/components/headers/ContentLength"
        Connection:
          $ref: "#/components/headers/Connection"
        Date:
          $ref: "#/components/headers/Date"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RequestTimeout:
      description: The request body could not be received before the timeout.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
        Content-Length:
          $ref: "#/components/headers/ContentLength"
        Connection:
          $ref: "#/components/headers/Connection"
        Date:
          $ref: "#/components/headers/Date"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RequestTooLarge:
      description: The framed request body exceeds the fixed limit.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
        Content-Length:
          $ref: "#/components/headers/ContentLength"
        Connection:
          $ref: "#/components/headers/Connection"
        Date:
          $ref: "#/components/headers/Date"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnsupportedMediaType:
      description: The request media type is unsupported.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
        Content-Length:
          $ref: "#/components/headers/ContentLength"
        Connection:
          $ref: "#/components/headers/Connection"
        Date:
          $ref: "#/components/headers/Date"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ExpectationFailed:
      description: The request expectation is unsupported.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
        Content-Length:
          $ref: "#/components/headers/ContentLength"
        Connection:
          $ref: "#/components/headers/Connection"
        Date:
          $ref: "#/components/headers/Date"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: An internal contract failure occurred.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
        Content-Length:
          $ref: "#/components/headers/ContentLength"
        Connection:
          $ref: "#/components/headers/Connection"
        Date:
          $ref: "#/components/headers/Date"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: A required service or dependency is not ready.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/ContentTypeOptions"
        Content-Length:
          $ref: "#/components/headers/ContentLength"
        Connection:
          $ref: "#/components/headers/Connection"
        Date:
          $ref: "#/components/headers/Date"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ProcessRequest:
      type: object
      additionalProperties: false
      required:
        - api_version
        - draft
        - message
        - smtp
      properties:
        api_version:
          $ref: "#/components/schemas/APIVersion"
        draft:
          $ref: "#/components/schemas/DraftVersion"
        message:
          $ref: "#/components/schemas/MessageInput"
        smtp:
          $ref: "#/components/schemas/SMTPInput"
        reporting:
          description: >
            Optional daemon reporting authority. When supplied for an accepting
            outcome, the daemon returns the exact Authentication-Results action;
            adapters must not synthesize it.
          $ref: "#/components/schemas/ReportingContext"
    SignRequest:
      type: object
      additionalProperties: false
      required:
        - api_version
        - draft
        - message
        - smtp
        - context
      properties:
        api_version:
          $ref: "#/components/schemas/APIVersion"
        draft:
          $ref: "#/components/schemas/DraftVersion"
        message:
          $ref: "#/components/schemas/MessageInput"
        smtp:
          $ref: "#/components/schemas/SMTPInput"
        context:
          $ref: "#/components/schemas/SigningContext"
    ReviseRequest:
      type: object
      additionalProperties: false
      required:
        - api_version
        - draft
        - message
        - smtp
        - incoming_smtp
        - context
      properties:
        api_version:
          $ref: "#/components/schemas/APIVersion"
        draft:
          $ref: "#/components/schemas/DraftVersion"
        message:
          $ref: "#/components/schemas/MessageInput"
        smtp:
          $ref: "#/components/schemas/SMTPInput"
        incoming_smtp:
          description: >
            Exact inherited envelope evidence verified before revision.
            The smtp member is the distinct outgoing signing envelope.
          $ref: "#/components/schemas/SMTPInput"
        context:
          $ref: "#/components/schemas/SigningContext"
    MessageInput:
      type: object
      additionalProperties: false
      required:
        - raw_rfc5322_base64
      properties:
        raw_rfc5322_base64:
          type: string
          format: byte
          minLength: 0
          maxLength: 44739244
        fidelity:
          type: string
          enum:
            - raw_rfc5322
            - milter_reconstructed_crlf
    SMTPInput:
      type: object
      additionalProperties: false
      required:
        - mail_from
        - rcpt_to
      properties:
        mail_from:
          type: string
          minLength: 0
          maxLength: 256
        rcpt_to:
          type: array
          minItems: 1
          maxItems: 2000
          uniqueItems: false
          items:
            type: string
            minLength: 0
            maxLength: 256
    ProcessResponse:
      type: object
      additionalProperties: false
      required:
        - api_version
        - draft
        - verification
        - policy
        - replay
        - disposition
        - actions
      properties:
        api_version:
          $ref: "#/components/schemas/APIVersion"
        draft:
          $ref: "#/components/schemas/DraftVersion"
        verification:
          $ref: "#/components/schemas/VerificationResult"
        policy:
          $ref: "#/components/schemas/PolicyResult"
        replay:
          $ref: "#/components/schemas/ReplayResult"
        disposition:
          $ref: "#/components/schemas/Disposition"
        actions:
          $ref: "#/components/schemas/ActionPlan"
    SigningContext:
      type: object
      additionalProperties: false
      required:
        - tenant
        - domain
      properties:
        tenant:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[a-z0-9][a-z0-9._-]{0,127}$"
        domain:
          type: string
          minLength: 1
          maxLength: 253
          pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?(?:\\.[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?)*$"
    ReportingContext:
      type: object
      additionalProperties: false
      required:
        - authserv_id
      properties:
        authserv_id:
          type: string
          minLength: 1
          maxLength: 253
          pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?(?:\\.[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?)*$"
    OperationResponse:
      type: object
      additionalProperties: false
      description: >
        Result and disposition are coherent: pass permits accept or continue,
        fail and permerror require reject, and temperror requires tempfail.
        Only accept may carry actions.
      required:
        - api_version
        - draft
        - operation
        - result
        - disposition
        - actions
      properties:
        api_version:
          $ref: "#/components/schemas/APIVersion"
        draft:
          $ref: "#/components/schemas/DraftVersion"
        operation:
          type: string
          enum:
            - sign
            - revise
        result:
          type: string
          enum:
            - pass
            - fail
            - permerror
            - temperror
        disposition:
          $ref: "#/components/schemas/Disposition"
        actions:
          $ref: "#/components/schemas/ActionPlan"
    Disposition:
      type: string
      enum:
        - accept
        - reject
        - tempfail
        - continue
    ActionPlan:
      type: array
      maxItems: 3
      items:
        $ref: "#/components/schemas/AddHeaderAction"
...
    VerificationResult:
      type: object
      additionalProperties: false
      required:
        - state
        - primary_reason
        - scope
        - historical_content
        - historical_signatures
        - custody_structure
        - checks
        - signature_sets
      properties:
        state:
          $ref: "#/components/schemas/VerificationState"
        primary_reason:
          $ref: "#/components/schemas/VerificationReason"
        scope:
          type: string
          enum:
            - current
        historical_content:
          type: string
          enum:
            - not_evaluated
        historical_signatures:
          type: string
          enum:
            - not_evaluated
        custody_structure:
          type: string
          enum:
            - not_evaluated
            - not_present
            - nd_links_evaluated
            - terminal_nd_requires_oob
        checks:
          type: array
          minItems: 1
          maxItems: 128
          items:
            $ref: "#/components/schemas/VerificationCheck"
        signature_sets:
          type: array
          maxItems: 16
          items:
            $ref: "#/components/schemas/SignatureSetResult"
        target:
          $ref: "#/components/schemas/VerificationTarget"
    VerificationTarget:
      type: object
      additionalProperties: false
      required:
        - sequence
        - instance
      properties:
        sequence:
          $ref: "#/components/schemas/CanonicalUint64"
        instance:
          $ref: "#/components/schemas/CanonicalUint64"
    VerificationCheck:
      type: object
      additionalProperties: false
      required:
        - class
        - reason
      properties:
        class:
          type: string
          enum:
            - message
            - protocol
            - body_hash
            - header_hash
            - signature
            - key
            - timestamp
            - envelope
            - domain_alignment
            - next_domain
            - provider
            - internal_contract
        reason:
          $ref: "#/components/schemas/VerificationReason"
    SignatureSetResult:
      type: object
      additionalProperties: false
      required:
        - algorithm
        - status
        - reason
        - key_policy
      properties:
        algorithm:
          type: string
          enum:
            - rsa-sha256
            - ed25519-sha256
            - unknown
        status:
          type: string
          enum:
            - pass
            - fail
            - permerror
            - temperror
            - ignored
        reason:
          $ref: "#/components/schemas/VerificationReason"
        key_policy:
          $ref: "#/components/schemas/KeyPolicyResult"
...
    PolicyResult:
      type: object
      additionalProperties: false
      required:
        - mode
        - verdict
        - primary_reason
        - do_not_modify
        - do_not_explode
        - dns_testing_effective
        - feedback
        - findings
      properties:
        mode:
          type: string
          enum:
            - strict
            - permissive
            - testing
        verdict:
          type: string
          enum:
            - accept
            - reject
            - tempfail
            - continue
        primary_reason:
          $ref: "#/components/schemas/PolicyReason"
        do_not_modify:
          type: string
          enum:
            - not_evaluated
        do_not_explode:
          type: string
          enum:
            - not_evaluated
        dns_testing_effective:
          type: boolean
        feedback:
          $ref: "#/components/schemas/PolicyFeedback"
        findings:
          type: array
          minItems: 1
          maxItems: 128
          items:
            $ref: "#/components/schemas/PolicyFinding"
...
    ReplayResult:
      type: object
      additionalProperties: false
      required:
        - class
      properties:
        class:
          type: string
          enum:
            - not_checked
            - disabled
            - first_seen
            - replayed
            - indeterminate
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - api_version
        - draft
        - code
        - category
      properties:
        api_version:
          $ref: "#/components/schemas/APIVersion"
        draft:
          $ref: "#/components/schemas/DraftVersion"
        code:
          type: string
          enum:
            - invalid_json
            - invalid_contract
            - unsupported_version
            - unsupported_draft
            - request_too_large
            - unsupported_media_type
            - not_found
            - method_not_allowed
            - forbidden
            - service_not_ready
            - service_overloaded
            - request_timeout
            - request_deadline
            - expectation_failed
            - precondition_failed
            - internal_error
        category:
          type: string
          enum:
            - request
            - availability
            - internal
    APIVersion:
      type: string
      enum:
        - v1
    DraftVersion:
      type: string
      enum:
        - draft-ietf-dkim-dkim2-spec-04
    CanonicalUint64:
      type: string
      pattern: "^[1-9][0-9]{0,19}$"
    VerificationState:
      type: string
      enum:
        - PASS
        - FAIL
        - PERMERROR
        - TEMPERROR
    VerificationReason:
      type: string
      enum:
        - none
        - limit_exceeded
        - malformed_message
        - malformed_protocol
        - missing_protocol
        - sequence_invalid
        - unsupported_algorithm
        - hash_mismatch
        - signature_mismatch
        - missing_key
        - invalid_key
        - ambiguous_key
        - revoked_key
        - unsupported_key_type
        - key_algorithm_mismatch
        - provider_temporary
        - provider_permanent
        - provider_contract
        - timestamp_invalid
        - envelope_mismatch
        - domain_alignment_mismatch
        - next_domain_mismatch
        - out_of_band_required
        - internal_contract
    PolicyReason:
      type: string
      enum:
        - protocol_pass
        - protocol_fail
        - protocol_permerror
        - protocol_temperror
        - permissive_override
        - testing_mode_observe
        - dns_testing_effective
        - dns_testing_mixed
        - dns_testing_ineligible
        - donotmodify_not_evaluated
        - donotexplode_not_evaluated
        - feedback_requested
        - feedback_relay_selected
        - feedhere_inert
        - exploded_reported